Time Passed, Global, Buildings
Added a way of measuring time passed (in timsTestPage).

Added the Globals file, which can keep track of global variables like for tracking how many buildings/resources the player as of what type using Maps.

The buildings file has a method for checking which buildings the player has and then methods for calculating what each building does in the given amount of time.