Added components file and edited main page
Components file added. Any widgets that we want to reuse, or could be useful in anyway and thus want to call multiple times should go in here. I have put the table, tablerows, and the navigation bar in here. Table rows are currently text, but can be changed to buttons if we want.